simulate: redesign to simulate each P
Currently, the simulator alternates a single "thread" of simulation
between a phase of shared mutator execution and background garbage
collection and a phase of mutator allocation and assist. As a result,
mutator assist takes 100% of the "CPU" and blocks out background
GC. If there's a significant amount of mutator assist, this can affect
the simulation results in significant ways.

Switch to simulating a set of Ps that each execute independently. Each
P does only one of background GC, idle GC, or mutator execution during
each schedule quantum, but can switch on quantum boundaries. This lets
us model the background GC scheduling decision, and allows for
overlapping mutator assists and background GC.

This does not interleave within a quantum, which has some
downsides. We set the quantum short enough and the heap high enough so
that we simulate enough schedule quanta to amortize the effect of
this, but it would be better to use something like a discrete event so
that the quantum only affects scheduling decisions and the not the
fidelity of the simulation itself.