Use a continuous time simulation
Currently, the simulation of each P is stepped forward by a scheduler
quantum. That means the Ps can get out of sync with each other by a
whole quantum, which means the fidelity of the simulation depends on
the scheduler quantum.

Fix this by instead tracking the time of the next event on each P and
simulating all Ps forward by that much time. Scheduler preemption is
now just another event and it can be desynchronized between the
Ps. Furthermore, we can compute the exact time that we'll run out of
work if that will happen before the next event and simulate up to that
time, rather than often ending at a ragged time when one P happens to
go over the work limit.